requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Migrate==4.0.4
python-dotenv==1.0.0
werkzeug==2.3.7
opencv-python==4.8.0.76
face-recognition==1.3.5
qrcode[pil]==7.4.2
Pillow==10.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
dlib==19.24.2
